+++ src/command_modules/azure-cli-sql/azure/cli/command_modules/sql/_validators.py
@@ -131,8 +131,8 @@ def validate_subnet(cmd, namespace):
 
 
 def validate_managed_instance_storage_size(cmd, namespace):
-    # Validate if entered storage size value is an increment of 32
-    if namespace.storage_size_in_gb and namespace.storage_size_in_gb % 32 == 0:
+    # Validate if entered storage size value is an increment of 32 if provided
+    if (not namespace.storage_size_in_gb) or (namespace.storage_size_in_gb and namespace.storage_size_in_gb % 32 == 0):
         pass
     else:
         raise CLIError('incorrect usage: --storage must be specified in increments of 32 GB')